pyproject.toml
[project]
name = "letterboxg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cloudscraper>=1.2.71",
    "google-genai>=1.68.0",
    "letterboxdpy>=6.5.0",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.33.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.8",
    "pytest>=8.0.0",
    "pytest-playwright>=0.4.0",
    "pytest-mock>=3.12.0",
]
